Fix hang related issue during installation
There are some fonts formula that might take little bit longer
time to download and installation and unfortunately this does not
give any feedback to the user about what's fontist doing under the
hood.

This commit adds an option to our DSL so, now we can tag these
formula so during installation it will show the progress bar.

This commit also updates the default font lookup locations for
Mac OS, so now fontist will also look into  windows related paths
before installation.

Fixes - 86
