+++ lib/fontist/utils.rb
@@ -1,4 +1,5 @@
 require "fontist/utils/dsl"
+require "fontist/utils/downloader"
 require "fontist/utils/zip_extractor"
 require "fontist/utils/exe_extractor"
